pyproject.toml
[project]
name = "library"
version = "0.1.0"
description = "Library Borrowing & Membership Management System"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apscheduler>=3.11.0",
    "argon2-cffi>=25.1.0",
    "email-validator>=2.2.0",
    "fastapi>=0.116.0",
    "jinja2>=3.1.6",
    "psycopg[binary]>=3.2.9",
    "pydantic>=2.11.0",
    "pyjwt>=2.10.0",
    "python-multipart>=0.0.20",
    "ruff>=0.15.10",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.35.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=8.4.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
